move auth related functions to auth.utils
