+++ backend/ops_api/ops/auth/utils.py
@@ -9,8 +9,12 @@ from authlib.jose import JsonWebToken
 from authlib.jose import jwt as jose_jwt
 from flask import current_app
 from marshmallow import ValidationError
+from sqlalchemy import select
 from sqlalchemy.exc import PendingRollbackError
+from sqlalchemy.orm import load_only
 
+from models import Role, User
+from ops_api.ops.auth.auth_types import UserInfoDict
 from ops_api.ops.auth.exceptions import NotActiveUserError
 from ops_api.ops.utils.response import make_response_with_headers
 
@@ -102,3 +106,42 @@ def decode_user(
     # claims = jwt.decode(payload, get_jwks(provider), claims_options=claims_options)
     claims = jwt.decode(payload, get_jwks(provider))
     return claims
+
+
+def register_user(userinfo: UserInfoDict) -> User:
+    user = get_user_from_token(userinfo)
+    if user:
+        return user, False
+    else:
+        # Create new user
+        # Default to an 'unassigned' role.
+        current_app.logger.debug("Creating new user")
+        try:
+            # Find the role with the matching permission
+            role = current_app.db_session.query(Role).options(load_only(Role.name)).filter_by(name="unassigned").one()
+            user = User(
+                email=userinfo["email"],
+                oidc_id=userinfo["sub"],
+                roles=[role],
+            )
+
+            current_app.db_session.add(user)
+            current_app.db_session.commit()
+            return user, True
+        except Exception as e:
+            current_app.logger.debug(f"User Creation Error: {e}")
+            current_app.db_session.rollback()
+            return None, False
+
+
+def get_user_from_token(userinfo: UserInfoDict) -> Optional[User]:
+    if userinfo is None:
+        return None
+    try:
+        stmt = select(User).where((User.oidc_id == userinfo["sub"]))
+        users = current_app.db_session.execute(stmt).all()
+        if users and len(users) == 1:
+            return users[0][0]
+    except Exception as e:
+        current_app.logger.debug(f"User Lookup Error: {e}")
+        return None
