+++ backend/ops_api/ops/resources/agreements.py
@@ -29,7 +29,7 @@ from models.cans import (
     ContractAgreement,
     ServiceRequirementType,
 )
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, OPSMethodView, handle_api_error
 from ops_api.ops.resources.agreements_constants import (
