+++ backend/ops_api/ops/resources/can_fiscal_year.py
@@ -5,7 +5,7 @@ from typing_extensions import override
 
 from models.base import BaseModel
 from models.cans import CANFiscalYear
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseListAPI, handle_api_error
 from ops_api.ops.utils.response import make_response_with_headers
