+++ backend/ops_api/ops/resources/cans.py
@@ -10,7 +10,7 @@ from typing_extensions import override
 
 from models.base import BaseModel
 from models.cans import CAN
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.utils.errors import error_simulator
