+++ backend/ops_api/ops/resources/portfolio_funding_summary.py
@@ -2,7 +2,7 @@ from flask import Response, request
 from typing_extensions import override
 
 from models.base import BaseModel
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, handle_api_error
 from ops_api.ops.utils.fiscal_year import get_current_fiscal_year
