+++ backend/ops_api/ops/resources/portfolio_status.py
@@ -3,7 +3,7 @@ from typing_extensions import override
 
 from models.base import BaseModel
 from models.portfolios import PortfolioStatus
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
