+++ backend/ops_api/ops/resources/procurement_steps.py
@@ -18,9 +18,10 @@ from models.workflows import (
     ProcurementStep,
     Solicitation,
 )
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.auth.exceptions import ExtraCheckError
+from ops_api.ops.auth.utils import get_user_from_token
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.schemas.procurement_steps import (
     AcquisitionPlanningRequest,
@@ -46,7 +47,6 @@ from ops_api.ops.utils.api_helpers import (
 )
 from ops_api.ops.utils.events import OpsEventHandler
 from ops_api.ops.utils.response import make_response_with_headers
-from ops_api.ops.utils.user import get_user_from_token
 
 
 def get_current_user_id():
