+++ backend/ops_api/ops/resources/research_projects.py
@@ -12,7 +12,7 @@ from models import CAN, Agreement, BudgetLineItem, MethodologyType, OpsEventType
 from models.base import BaseModel
 from models.cans import CANFiscalYear
 from models.projects import ResearchProject
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.utils.events import OpsEventHandler
