+++ backend/ops_api/ops/resources/users.py
@@ -6,7 +6,7 @@ from marshmallow import Schema
 from typing_extensions import override
 
 from models import BaseModel, User
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.schemas.users import PATCHRequestBody, POSTRequestBody, QueryParameters, UserResponse
