+++ backend/ops_api/ops/resources/workflow_submit.py
@@ -19,12 +19,12 @@ from models.workflows import (
     WorkflowStepStatus,
     WorkflowTriggerType,
 )
-from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.auth_types import Permission, PermissionType
 from ops_api.ops.auth.decorators import is_authorized
+from ops_api.ops.auth.utils import get_user_from_token
 from ops_api.ops.base_views import BaseItemAPI, handle_api_error
 from ops_api.ops.schemas.budget_line_items import PATCHRequestBody
 from ops_api.ops.utils.response import make_response_with_headers
-from ops_api.ops.utils.user import get_user_from_token
 
 ENDPOINT_STRING = "/workflow-submit"
